requirements.txt
tensorflow>=2.12.0
opencv-python>=4.8.0
numpy>=1.24.0
protobuf>=3.20.0
openai>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
